Update to 1.99.1 - drop sarea and perform swap buffer in X server.
Still to resolve is the swap buffer request.  It should probably be
broken into two requests, one to post the swap request and one to wait
for it to be completed.  Also, need to find a good solution to
CopySubBuffer that doesn't require a roundtrip per rectangle.

Don't need to solve all this for 2.0, though, can add requests later on.